File header, pseudo, struct
Created file header, added the #include directives, #defined a variable for maximum chars for input. Then created a typedef struct, within the struct are char pointers for the destination and date. Since I am starting this project a little later after the first lesson I am a bit rusty with malloc so I may have to go back and ensure I passed the correct argument to it (number of bytes) by using the variable I created. Then in main I started to pseudocode the requirements for taking user input.