Function Creation Requirements + Other Reqs
Went through function creation requirements and created the array variable of struct, the function prototype of fillFlightInfo and printFlightInfo, as well as an outline of their definitions. For these definitions I basically just added the headers and notes within the body that are either from requirements or just things I may have thought of while thinking about how the function should work.