Relocated Malloc to Function
After getting an error from the compiler I realized that I could not use malloc within the struct (unless i am still wrong), so I moved declaring the pointers to malloc to the fillFlightInfo function where it does say in the requirements to do so, i was previously under the impression that we had to do it in both places. I also just remembered you have to error check when using malloc and, in c, you cannot do this in a struct. I also quickly added strcpy because my next step is to create the portion in main which takes user input and calls the fillFlightInfo function to ensure that I am properly using malloc with the struct.