printFlightInfo
I couldn't fall asleep knowing that I was almost done this project so I decided to keep going. I finished the function definition for printFlightInfo. I feel like this was a trick question because I wanted to add a comma or other words but it clearly says in the first (and next) 35 characters of the LINE... so to be safe i won't add anything else for now. I cleaned the project up a bit by removing some notes-to-self that I rectified. Also ensured to free the space up, however I may have to revisit it to ensure that I have done this correctly. Added some questions at the top of the page to ask prof in person. also changed the file header a bit.