Free()
Created a for loop to iterate through each element in the array to free the memory that was allocated in the fillFlightInfo function, previously I was simply freeing memory of the first index. Cleaned it up a bit by removing some notes to self and added a couple more comments for clarification. Also removed some redundant pieces of code (like taking the return of fgets to have the option of sending it to the fillFlightsInfo function.